Re-request data when dates change
Refactored data access, top destinations and table to be classes so they
can be referred to as instances with private state.
This allows to re-request data and update the private state of the
components rather then calling a function that does everything at once.
Main is updates to take advantage of this.